backend/requirements.txt
# Django本体
django

# APIを作るための、事実上の標準ライブラリ
djangorestframework

# フロントエンドとの通信を許可するため
django-cors-headers

# Djangoの認証を強化するためのライブラリ
clerk-django
requests

# PostgreSQLとDjangoを繋ぐための部品
psycopg2-binary

# WebSocket(Django Channels)を動かすための部品
channels
channels-redis # 本番環境ではこちらを推奨しますが、開発ではインメモリでOK
daphne # ASGIサーバー
uvicorn
websockets

gunicorn
dj-database-url

whitenoise[brotli]

svix

cryptography

# Django Channlesみたいなやつ入れてないけど大丈夫なのかな